Moved decal creation to MeshCreator.
