+++ Assets/GothicVR/Scripts/Creator/VobCreator.cs
@@ -24,8 +24,6 @@ namespace GVR.Creator
 {
 	public class VobCreator : SingletonBehaviour<VobCreator>
     {
-        private const float decalOpacity = 0.5f;
-        
         private MeshCreator meshCreator;
         private SoundCreator soundCreator;
         private AssetCache assetCache;
@@ -204,7 +202,7 @@ namespace GVR.Creator
         {
             var spot = GameObject.CreatePrimitive(PrimitiveType.Sphere);
             Destroy(spot.GetComponent<SphereCollider>()); // No need for collider here!
-
+            
             if (DebugSettings.Instance.EnableVobFPMesh)
             {
 #if UNITY_EDITOR
@@ -217,6 +215,8 @@ namespace GVR.Creator
             }
             else
             {
+                // Quick win: If we don't want to render the spots, we just remove the Renderer.
+                // FIXME - Loading can be optimized with a proper Prefab
                 Destroy(spot.GetComponent<MeshRenderer>());
             }
 
@@ -236,35 +236,7 @@ namespace GVR.Creator
         {
             var parent = parentGos[vob.type];
             
-            if (vob.vobDecal.HasValue)
-            {
-                var decalData = vob.vobDecal.Value;
-
-                var decalProjectorGo = new GameObject(decalData.name);
-                var decalProj = decalProjectorGo.AddComponent<DecalProjector>();
-                var texture = assetCache.TryGetTexture(vob.visualName);
-                
-                // x/y needs to be made twice the size and transformed from cm in m.
-                // z - value is close to what we see in Gothic spacer.
-                decalProj.size = new(decalData.dimension.X * 2 / 100, decalData.dimension.Y * 2 / 100, 0.5f);
-                decalProjectorGo.SetParent(parent);
-                SetPosAndRot(decalProjectorGo, vob.position, vob.rotation!.Value);
-                
-                decalProj.pivot = UnityEngine.Vector3.zero;
-                decalProj.fadeFactor = decalOpacity;
-                
-                // FIXME use Prefab!
-                // https://docs.unity3d.com/Packages/com.unity.render-pipelines.high-definition@12.0/manual/creating-a-decal-projector-at-runtime.html
-                var standardShader = Shader.Find("Shader Graphs/Decal");
-                var material = new Material(standardShader);
-                material.SetTexture(Shader.PropertyToID("Base_Map"), texture);
-
-                decalProj.material = material;
-
-                return;
-            }
-
-            Debug.LogWarning("No decalData was set for: " + vob.visualName);
+            meshCreator.CreateDecal(vob, parent);
         }
         
         private GameObject CreateDefaultMesh(PxVobData vob)
